fix(components): add total value for contract detail
