+++ src/lib/pages/account-details/components/delegations/index.tsx
@@ -5,7 +5,7 @@ import { useEffect } from "react";
 import { AmpEvent, track } from "lib/amplitude";
 import { Loading } from "lib/components/Loading";
 import { ErrorFetching } from "lib/components/state";
-import { useUserDelegationInfos } from "lib/pages/account-details/data";
+import { useAccountDelegationInfos } from "lib/model/account";
 import type { BechAddr } from "lib/types";
 import { getTokenLabel } from "lib/utils";
 
@@ -38,7 +38,7 @@ export const DelegationsSection = ({
     rewards,
     redelegations,
     totalCommission,
-  } = useUserDelegationInfos(address);
+  } = useAccountDelegationInfos(address);
 
   useEffect(() => {
     onClose();
